repo: Move fsverity bits to ostree-repo-verity.c
This file will get larger when we start doing more with fsverity.
